Move mobilenet model definitions to discojs/src/models folder
