+++ .gitignore
@@ -1,7 +1,8 @@
 # dependencies
 /node_modules/
-# disco models
-**/models
+# model.json files
+server/models
+cli/models
 # built
 dist/
