+++ discojs/src/default_tasks/index.ts
@@ -1,7 +1,7 @@
-export { cifar10 } from './cifar10/index.js'
+export { cifar10 } from './cifar10.js'
 export { lusCovid } from './lus_covid.js'
 export { skinMnist } from './skin_mnist.js'
 export { mnist } from './mnist.js'
-export { simpleFace } from './simple_face/index.js'
+export { simpleFace } from './simple_face.js'
 export { titanic } from './titanic.js'
 export { wikitext } from './wikitext.js'
